+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## v1.0.2 (2025-06-13)
+
+- npm install sweetalert2
+- src/app.html => ambil head dari web-template-contact-management/register.html
+- src/lib/alert.js
+- src/lib/api/UserApi.js
+- src/routes/(default)/+layout.svelte => dari <body class="flex items-center justify-center min-h-screen p-4 bg-gradient-to-br from-gray-900 to-gray-800"> ubah jadi div
+- src/routes/(default)/register/+page.svelte => isi dibawah body register.html
+
 ## v1.0.1 (2025-06-09)
 
 - disable SSR (Server Side Rendering)